Multilanguage site + locale
When you have a multilanguage site you should always add the language
to your url. Unless the language is dependent of the hostname, but this
isn't realy possible in the current symfony version unless you do some
hacking.

So we don't specify the defaultLocale in the SlugRouter, else when you
generate an url for your default locale it will just be
http://hostname/ while it should be http://hostname/defaultlocale … and
then you also have two urls for the same content. This can also lead to
problems with a language chooser (if you have one)